fix google analytics
