+++ src/pages/_app.tsx
@@ -3,28 +3,50 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import "react-activity/dist/Dots.css";
 import "@/styles/globals.css";
-import { useEffect } from "react"
+import { useEffect } from "react";
 import type { AppProps } from "next/app";
-import { useRouter } from "next/router"
-import { EnvironmentHelper, pageview } from "@/utils";
+import Head from "next/head";
+import { useRouter } from "next/router";
+import { EnvironmentHelper, GoogleAnalyticsHelper } from "@/utils";
 import { AuthProvider } from "@/hooks/useAuth";
 
 EnvironmentHelper.init();
 
 function MyApp({ Component, pageProps }: AppProps) {
-  const router = useRouter()
+  const router = useRouter();
   useEffect(() => {
-    const handleRouteChange = (url: string) => {
-      pageview(url)
+    if (EnvironmentHelper.GoogleAnalyticsTag) {
+      const handleRouteChange = (url: string) => {
+        GoogleAnalyticsHelper.pageview(url);
+      };
+      router.events.on("routeChangeComplete", handleRouteChange);
+      return () => {
+        router.events.off("routeChangeComplete", handleRouteChange);
+      };
     }
-    router.events.on('routeChangeComplete', handleRouteChange)
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
+  }, [router.events]);
 
   return (
     <AuthProvider>
+      <Head>
+        {EnvironmentHelper.GoogleAnalyticsTag && (
+          <>
+            <script async src={`https://www.googletagmanager.com/gtag/js?id=${EnvironmentHelper.GoogleAnalyticsTag}`} />
+            <script
+              dangerouslySetInnerHTML={{
+                __html: `
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', '${EnvironmentHelper.GoogleAnalyticsTag}', {
+              page_path: window.location.pathname,
+            });
+          `,
+              }}
+            />
+          </>
+        )}
+      </Head>
       <Component {...pageProps} />
     </AuthProvider>
   );
