+++ src/pages/_document.tsx
@@ -1,5 +1,4 @@
 import Document, { Html, Head, Main, NextScript } from "next/document";
-import { EnvironmentHelper } from "@/utils"
 
 class MyDocument extends Document {
   render() {
@@ -9,19 +8,6 @@ class MyDocument extends Document {
           <link rel="icon" href="/favicon.ico" />
           <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet" />
           <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.2/css/all.css" integrity="sha384-oS3vJWv+0UjzBfQzYUhtDYW+Pj2yciDJxpsK1OYPAYjqT085Qq/1cq5FLXAZQ7Ay" crossOrigin="anonymous" />
-          <script async src={`https://www.googletagmanager.com/gtag/js?id=${EnvironmentHelper.GoogleAnalyticsTag}`} />
-          <script
-            dangerouslySetInnerHTML={{
-              __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${EnvironmentHelper.GoogleAnalyticsTag}', {
-              page_path: window.location.pathname,
-            });
-          `,
-            }}
-          />
         </Head>
         <body>
           <Main />
