+++ src/utils/index.ts
@@ -1,6 +1,6 @@
 export { ApiHelper, ArrayHelper, UserHelper } from "@/appBase/helpers";
 export { EnvironmentHelper } from "./EnvironmentHelper";
-export { Permissions } from "./Permissions"
+export { Permissions } from "./Permissions";
 
 export * from "./interfaces";
-export * from "./googleAnalyticsHelper"
+export { GoogleAnalyticsHelper } from "./GoogleAnalyticsHelper";
